refactor: Some React-style wiring
